fix(wave4): adversarial-review findings in the Phase 4 leftovers Rust
- manufacturing_tolerances.rs: analyze_clearance etch fallback 0.06 ->
  the oracle's 0.05 (the shipped mutant — VERIFICATION.md's 'caught'
  claim was false; the clearance-side fallback case in the RED commit
  discriminates it). FeatureTolerance.nominal_value and (clearance)
  worst_case_max now carry the ORIGINAL caller object (Py<PyAny>) like
  the monte_carlo dataclasses, so an int clearance stays int: repr '1'
  not '1.0', dataclass == via CPython's own == on the preserved fields,
  repr via CPython repr. from_py_object dropped (requires Clone, which
  Py<PyAny> cannot provide; nothing extracts the type from an argument).
- hypergraph_factory.rs: the global-net filter compares in i64
  ((n_pins as i64) > threshold) so a negative threshold filters every
  net — the 'as usize' wrap filtered none.
- manufacturing_monte_carlo.rs: inner-dimension parity for positions and
  bounds, replicating numpy's exact failure classes at the oracle's op
  order: positions trailing dim 1 broadcasts (x -> [x, x]), dim 0 or >=3
  raises numpy's broadcast ValueError (exact text incl. trailing space),
  bounds trailing dim 0/1 raise numpy's IndexError text; bounds >=3
  tolerated (oracle indexes only columns 0/1). No more silent compute on
  (N,3) positions and no PanicException for (N,1) bounds.
- stackup_validator.rs (temper-io-types): validate_stackup keeps the
  ORIGINAL impedance_spec_ohms object; check_impedance_spec renders the
  messages from it via CPython str() (int 90 -> '90 Omega', float 90.0
  -> '90.0 Omega') while the branch comparisons use the extracted f64.

GREEN: 133 differential + PBT suites, cargo test (53 + 37 Rust tests),
clippy clean on both crates.

+++ packages/temper-design-bundle/src/hypergraph_factory.rs
@@ -168,7 +168,11 @@ impl HypergraphFactory {
         let mut valid_indices: Vec<usize> = Vec::new();
         for (i, net) in net_list.iter().enumerate() {
             let n_pins = net.getattr("pins")?.len()?;
-            if self.ignore_global_nets && n_pins > self.global_net_threshold as usize {
+            // `len(net.pins) > global_net_threshold` — the comparison stays
+            // in i64 so a NEGATIVE threshold filters every net (the oracle's
+            // `len > -5` is true for any len). Casting `-5i64 as usize`
+            // would wrap to a huge threshold and filter nothing.
+            if self.ignore_global_nets && (n_pins as i64) > self.global_net_threshold {
                 continue;
             }
             if n_pins >= 2 {
